@@ -2,26 +2,15 @@
   "name": "ensure-gitignore",
   "version": "1.1.3",
   "description": "Ensure the presence of patterns within a project's gitignore",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
+    "build": "tsc --project tsconfig.build.json",
     "lint": "eslint .",
     "release": "changeset publish",
     "test": "jest"
   },
   "jest": {
     "testEnvironment": "node"
-  },
-  "husky": {
-    "hooks": {
-      "commit-msg": "commitlint --edit --extends seek",
-      "pre-commit": "lint-staged"
-    }
-  },
-  "lint-staged": {
-    "*.js": [
-      "prettier --write",
-      "git add"
-    ]
   },
   "repository": {
     "type": "git",
@@ -35,14 +24,14 @@
   "homepage": "https://github.com/seek-oss/ensure-gitignore#readme",
   "devDependencies": {
     "@changesets/cli": "^2.17.0",
-    "eslint": "^5.7.0",
-    "eslint-config-seek": "^3.2.1",
-    "husky": "^1.1.2",
-    "jest": "^23.6.0",
-    "lint-staged": "^7.3.0",
-    "prettier": "^1.14.3"
+    "@types/jest": "^27.0.1",
+    "eslint": "^7.32.0",
+    "eslint-config-seek": "^7.0.8",
+    "eslint-config-skuba": "^1.0.11",
+    "jest": "^27.1.0",
+    "prettier": "^2.4.0"
   },
-  "release": {
-    "success": false
+  "dependencies": {
+    "typescript": "^4.4.2"
   }
 }
